feat(rentals): add the rentals list, insertion, update, and delete features
To list  all the rentals send a requisition GET to /rentals. To insert a new open rent in the
database send a requisition POST to /rentals with the information an body wich contains the an
object like {customerId, gameId, daysRented}, to update a closed rent send a requisition POST to
/rentals/id/return with where id is the id of the required rent. To delete a closed rent send a
requisition DELETE to /rentals/id with the id of the rent.